requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
face-recognition==1.3.0
dlib==19.24.1
pyzbar==0.1.9
Pillow==12.1.0
